Shrikrushna_214_BtechB_AI_DRIVEN_SMART_SEEDING/Project/ml-backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-pymongo==2.3.0
pandas==2.1.3
xgboost==2.0.2
scikit-learn==1.3.2
requests==2.31.0
google-cloud-texttospeech==2.14.1
google-cloud-vision==3.4.5
pillow==10.1.0
